requirements.txt
langchain
python-dotenv
ipykernel
langchain-community
pypdf
bs4
arxiv
pymupdf
fitz
frontend
wikipedia
langchain-text-splitters
langchain-openai
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
beautifulsoup4
streamlit
langchain_groq
langchain_core
fastapi
uvicorn
langserve
sse_starlette
langchain_community